Move DupConf to utils package.
